fix: select-mode marks are message ids, immune to list reprojection
msg_marks held indices into App::messages, but any re-read reprojects
that list — and a re-read can be triggered remotely (another
participant's edit/delete/reaction arrives over api-listen). Sequence:
mark messages 5-8 for deletion, someone deletes an earlier message,
the push-driven reload shifts every index, Shift+X deletes the wrong
messages. The same flaw in miniature: the cursor was clamped into
range after a reload instead of re-anchored.

Marks are now message ids (the same reasoning InFlight::MarkRead
already applied: carry the id, not the index). The reload handler
prunes marks whose message vanished and re-seeks the cursor's message
by id, falling back to the clamp only when it's gone. Regression test
covers both the survival and the pruning.

+++ src/tui/app.rs
@@ -593,9 +593,12 @@ pub struct App {
     /// popup's cancel path returns the user to Compose instead of
     /// leaving them stuck in Select mode.
     pub select_from_compose: bool,
-    /// Messages marked in Select mode for a multi-select action (indices into
-    /// [`Self::messages`]). Empty = the action falls back to the cursor message.
-    pub msg_marks: HashSet<usize>,
+    /// Messages marked in Select mode for a multi-select action, by
+    /// **message id** — an id survives the re-read that reprojects the list
+    /// (a remote edit/delete shifts every index), so a batch action can
+    /// never land on the wrong message. Empty = the action falls back to
+    /// the cursor message.
+    pub msg_marks: HashSet<u64>,
     /// A **sequential** batch of per-message ops (delete / react) over the
     /// multi-selection, in progress. The worker is serial, so the batch fires
     /// one request at a time — each response advances to the next — instead of
